move tremor into its own package
