+++ format/it/layout/channel/memory.go
@@ -3,10 +3,10 @@ package channel
 import (
 	"github.com/gotracker/voice/oscillator"
 
-	"github.com/gotracker/playback/format/internal/effect"
 	"github.com/gotracker/playback/format/internal/memory"
 	formatutil "github.com/gotracker/playback/format/internal/util"
 	oscillatorImpl "github.com/gotracker/playback/oscillator"
+	"github.com/gotracker/playback/tremor"
 )
 
 type SharedMemory struct {
@@ -46,7 +46,7 @@ type Memory struct {
 	panbrello          memory.Value[DataEffect] `usage:"Yxy"`
 	volChanVolumeSlide memory.Value[DataEffect] `usage:"vDxy"`
 
-	tremorMem           effect.Tremor
+	tremorMem           tremor.Tremor
 	vibratoOscillator   oscillator.Oscillator
 	tremoloOscillator   oscillator.Oscillator
 	panbrelloOscillator oscillator.Oscillator
@@ -155,7 +155,7 @@ func (m *Memory) Panbrello(input DataEffect) DataEffect {
 }
 
 // TremorMem returns the Tremor object
-func (m *Memory) TremorMem() *effect.Tremor {
+func (m *Memory) TremorMem() *tremor.Tremor {
 	return &m.tremorMem
 }
