+++ format/s3m/layout/channel/memory.go
@@ -3,10 +3,10 @@ package channel
 import (
 	"github.com/gotracker/voice/oscillator"
 
-	"github.com/gotracker/playback/format/internal/effect"
 	"github.com/gotracker/playback/format/internal/memory"
 	formatutil "github.com/gotracker/playback/format/internal/util"
 	oscillatorImpl "github.com/gotracker/playback/oscillator"
+	"github.com/gotracker/playback/tremor"
 )
 
 type SharedMemory struct {
@@ -36,7 +36,7 @@ type Memory struct {
 	tempoIncrease memory.Value[DataEffect]
 	lastNonZero   memory.Value[DataEffect]
 
-	tremorMem         effect.Tremor
+	tremorMem         tremor.Tremor
 	vibratoOscillator oscillator.Oscillator
 	tremoloOscillator oscillator.Oscillator
 	patternLoop       formatutil.PatternLoop
@@ -98,7 +98,7 @@ func (m *Memory) LastNonZeroXY(input DataEffect) (DataEffect, DataEffect) {
 }
 
 // TremorMem returns the Tremor object
-func (m *Memory) TremorMem() *effect.Tremor {
+func (m *Memory) TremorMem() *tremor.Tremor {
 	return &m.tremorMem
 }
