+++ tremor/tremor.go
@@ -1,4 +1,4 @@
-package effect
+package tremor
 
 // Tremor is the storage object for the tremor values
 type Tremor struct {
